requirements.txt
blis==0.7.7
catalogue==2.0.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cymem==2.0.6
es-core-news-md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-3.3.0/es_core_news_md-3.3.0-py3-none-any.whl
idna==3.3
install==1.3.5
Jinja2==3.1.2
langcodes==3.3.0
lxml==4.8.0
MarkupSafe==2.1.1
murmurhash==1.0.7
numpy==1.22.4
packaging==21.3
pathy==0.6.1
Pillow==9.1.1
preshed==3.0.6
pydantic==1.8.2
Pygments==2.12.0
pyparsing==3.0.9
pyspellchecker==0.6.3
requests==2.27.1
smart-open==5.2.1
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
thinc==8.0.16
tqdm==4.64.0
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
wasabi==0.9.1
